pkg/helm/release: move ManagerFactory to separate file
